refactor(protocol): deduplicate now_kerberos/time_diff, fix TGS clock skew
- Move now_kerberos() and time_diff() to validate module (shared by AS/TGS)
- Fix TGS-REP clock skew: only reject starttime/authtime too far in the future
  (authtime can legitimately be hours in the past for TGS exchanges)
- Remove duplicate UTC_OFFSET constant
